fixed kings' moveability

+++ piece/king.py
@@ -17,31 +17,35 @@ class King(Piece):
         return [(1+self._x, 0+self._y), (-1+self._x, 0+self._y), (0+self._x, 1+self._y), (0+self._x, -1+self._y),
                 (1+self._x, 1+self._y), (-1+self._x, -1+self._y), (1+self._x, -1+self._y), (-1+self._x, 1+self._y)]
 
-    def getMoveablePositions(self, recalculate:bool = False) -> list[tuple[int, int]]:
+    def getMoveablePositions(self, recalculate:bool = False, _list: list[tuple[int, int]] = None) -> list[tuple[int, int]]:
         if not recalculate:
             return self._moveablePositions
-        
-        self._moveablePositions = coords = self._getAdjacentCoords()
+
+        if _list == None:
+            self._moveablePositions = coords = self._getAdjacentCoords()
+        else:
+            coords = _list
+
         originalX = self._x
         originalY = self._y
         original = self._board.deletePieceFromTable(originalX, originalY)
         for i in range(len(coords)-1, -1, -1):
-            coord = coords[i]
-            x, y = coord
+            x, y = coords[i]
             isKingAddedToBoard = False
             deletedPiece = None
             # adds the kings to empty cells and checks if it can take off the enemy's pieces
             if self._board.isValidCoordinate(x, y):
                 if self._board.isTableCellEmpty(x, y):
-                    self._board.addPieceToTable(King(self._board, x, y, self._player, self._image), x, y, modifyCoordsInPiece=False)
+                    self._board.addPieceToTable(King(self._board, x, y, self._player, self._image), x, y, testMove=True)
                     isKingAddedToBoard = True
-                elif not self.isOwnedBySamePlayer(self._board.getBoardPiece(x, y)):
+                elif not self.isOwnedBySamePlayer(self._board.getBoardPiece(x, y)) and _list == None:
                     deletedPiece = self._board.deletePieceFromTable(x, y)
-                    canMove = self.isMoveable(x, y, recalculate=True)
-                    self._board.addPieceToTable(deletedPiece, x, y, modifyCoordsInPiece=False)
+                    canMove = (x, y) in self.getMoveablePositions(recalculate=True, _list=coords.copy())
+                    self._board.addPieceToTable(deletedPiece, x, y, testMove=True)
+
                     if not canMove:
                         del coords[i]
-                        break
+                        continue
             else:
                 del coords[i]
                 continue
@@ -66,7 +70,7 @@ class King(Piece):
             if isKingAddedToBoard:
                 self._board.deletePieceFromTable(x, y)
 
-        self._board.addPieceToTable(original, originalX, originalY, modifyCoordsInPiece=False)
+        self._board.addPieceToTable(original, originalX, originalY, testMove=True)
         return coords
 
 class BlackKing(King):
